source/s3-lambda-notification/requirements.txt
boto3
dnspython
aws_lambda_powertools